feat(deploy): add end-to-end testing and configuration updates
This commit introduces a comprehensive end-to-end testing framework for the trading system, ensuring that various trading scenarios are validated against expected outcomes. Additionally, it updates the Docker configuration to allow for circular references in Spring, enhancing the application's flexibility.

The changes include new scripts for testing, modifications to the aggregation service, and database migrations to support new data structures. These updates improve the robustness of the trading system while minimizing the risk of regressions in future deployments.

+++ services/market-data-consumer/src/main/java/com/tradestream/market_data_consumer/service/AggregationService.java
@@ -40,8 +40,8 @@ public class AggregationService {
         if (inserted == 0) return false; // duplicate
 
         String ticker = t.ticker().toUpperCase(Locale.ROOT);
-        BigDecimal price = BigDecimal.valueOf(t.price());
-        BigDecimal qty   = BigDecimal.valueOf(t.quantity());
+        BigDecimal price = t.price();
+        BigDecimal qty   = t.quantity();
 
         for (Interval itv : SUPPORTED) {
             Instant bucket = Bucketizer.bucketStart(t.timestamp(), itv);
